test/blobstore: allow variable number of io_units
Size of io_unit is tied to the block size of the underlying device, thus the
total number of io_units. It just so happened that when this test last executed
it was on a 512B device - when SPDK-CI now has 4KiB.

Other instances in the match file already allow for a variable number of
io unit size.

Change-Id: I6f78619503723670a0f593efbf26511335646869
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26548
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>