memory: keep (un)registerring when notify_cb returns error
There are cases that a vaddr is valid for some mappings but invalid for
others. For exmaple, a 4-KiB aligned vaddr may be valid for RDMA MR
mapping but not valid for vtophys mapping. In these cases, we should
allow the vaddrs to be (un)registered from/to any mappings that accept
them.

Change-Id: I23280ecf402ba7a6270d130a3107773ba42bc087
Signed-off-by: Jinlong Chen <chenjinlong.cjl@alibaba-inc.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26565
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>