nvmf: validate NSID for Identify CNS 03h command
According to NVMe specification r1.4, section 5.15.2.4, the controller
shall return the descriptor list only if the NSID field specifies an
active namespace. If the NSID is inactive, the controller shall abort
the command with the status Invalid Field in Command. If the NSID is
invalid, the controller shall abort the command with the status Invalid
Namespace or Format.

Change-Id: I88212b1a3f133855afeb59243fa4c0f5f2de0767
Signed-off-by: Dor Deri <dor.deri@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26583
Community-CI: Mellanox Build Bot
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>