lib/nvmf: avoid aborting reservation commands
We don't have support to abort namespace persistent reservation
commands, so add handling to identify if the cid refers to a reservation
command and return early as not aborted before falling into the
transport level and bdev level abort logic.

Add unit test to coverage identifying a cid as a reservation command.

Change-Id: If929d87a65bce77fcb43924ec4ab9eba5e80acd1
Signed-off-by: Joel Cunningham <joel.cunningham@oracle.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26597
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>