lib/nvmf: Add support custom discovery filter
This patch adds support for custom discovery filter for the
nvmf target.
This helps with use cases like segmented networks, where the
target may want to return only a subset of listeners based
on the network address on which the discovery request was
received.
This patch only adds the variable and exposes an API to set it.
Its usage in discovery log generation is a separate patch.

Change-Id: Ice02e5e8b58b33e14bba147339999f0fc5a97869
Signed-off-by: Abhineet Pandey <abhineet.pandey@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26609
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>