nvme/rdma: Abort outstanding accel transfer on qp disconnect
When we don't use SRQ and a qp is disconnected, we reap completions for
every submiited recv wr since they are generated by HW and not by remote
target. But when we use SRQ, we can reap recv completion only when a
target sends a response. In some cases (e.g. in test mlx5_driver_mkey)
target is dead and doesn't send anything and we can't comlete accel
requests which are waiting for the data transfer completion. To fix this
issue, manually complete data transfer (if any) when a qpair is
disconnected or in lingering state.

Change-Id: I4a315621c43f6fa55e632315cbe1d2a614e77909
Signed-off-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26914
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>