nvme/rdma: Fix group pointer in qpair_release_poller
The user can destroy IO qpair while it is connected state, in
that case generic nvme layer first disconnects qpair from the poll
group, that clears qpair->group pointer and then destoys the qpair in
transport-specific function.
In rdma level qpair has a poller which depends on the poll group and
keeps internal pointer to a poll group. We should use this pointer to
release the poller independetly on the qpair->group pointer.

Change-Id: I3645e05fa0c17dba0ee8e9f33ead365cdb287f03
Signed-off-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26916
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>