bdev/nvme: simplify unregister flow by using ref
The idea is to acquire a controller reference before the asynchronous
operation and release it upon completion.

With this change, the predicate + unregister pattern is no longer needed
in a few flows, and the code is simplified - nvme_ctrlr_unregister is
invoked only from two contexts now.

Reset flows still use the predicate + unregister pattern, as applying
reference get/put there is not trivial.

Change-Id: Ibac292bf562c7c3089c41260e85cf0c1cd32036f
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26754
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>