bdev/nvme: replace nvme_ctrlr_get_[first|next]_active_ns
These utilities are confusing, as there is an SPDK NVMe API that
iterates over namespaces rather than over the namespace list in
the bdev module.

Moreover, using foreach is simply more readable (especially compared
to a while loop), and _active_ as part of the function name was also
misleading, as namespaces on that list might not be active.

Change-Id: I27f29f184cd323c5d0b9ab8d063e4d0944a3b932
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26852
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Krzysztof Goreczny <krzysztof.goreczny@dell.com>