test/vfio-user: bump QEMU version for vfio-user to latest upstream release
Starting with QEMU v10.1.1 vfio-user-pci devices are fully supported without
any need for additional patches. This commit switches the repo to upstream and
branch to the latest version.

Note: Series starting with this patch was tested on my local Fedora 42 system,
and by the end result in passing score for SPDK_TEST_VFIOUSER/_QEMU test flags.
SPDK-CI automation is work in progress.

Change-Id: I945a7a4daa819a13419ed80e1122e1d12ad69507
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26857
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>