bdev/nvme: add util to unregister poller and continue reset
Prepare a utility that unregisters the connect poller and invokes
the reset iteration continuation. With this change, the code is
unified and the poller is unregistered in a single context.

There is a behavioral change in bdev_nvme_destroy_ctrlr_channel_cb,
as the poller is now unregistered only when reset_iter is set.
However, these are always set together, and if reset_iter was set
due to a destroy operation, unregistering the poller with NULL is
safe by design.

Change-Id: I8ccce5ba0a11453ad638e3cefb2e501391157b05
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26968
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Krzysztof Goreczny <krzysztof.goreczny@dell.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>