bdev/nvme: remove g_bdev_nvme_init_thread
It was introduced a while ago
commit 5f9bdac00, bdev/nvme: add RPC to enable and disable hotplug
as part of a new RPC method. However, since then it has remained the
only one that relies on the init thread — it should by design match
the SPDK app thread.

By removing it, the code is significantly simplified, and reasoning
about threads in the bdev/nvme module becomes easier.

Change-Id: I0d1e3dfd48768e98bd05999917557cb3e62c62a7
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27070
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Ben Walker <ben@nvidia.com>