bdev/nvme: remove thread from nvme_ctrlr
It is the app thread by design.

By keeping the number of threads to a minimum, it becomes much easier to
reason about thread synchronization - the code reader doesn’t need to
consider whether the controller thread matches the app thread or if the
two could race with each other.

While here, also fix bdev_nvme_create_qpair and getting the thread id.
Most likely, it should be the id of the thread where the qpair is
created, not the app thread.

Change-Id: Iaa5fa55f8d6ef500e48c25abadbfae9cf592fc12
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27072
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>