bdev/nvme: remove spdk_thread_exec_msg usage
In the set keys flow, it runs on the app thread by design, so there is
no need for exec_msg. The same applies to the firmware update flow.

For ctrlr put ref, it is actually better to have consistent asynchronous
behavior, so it was changed from exec_msg to send_msg.

Change-Id: Ie5a3efb7a514edfd7cd2fa09dadd61d9c9c237ec
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27074
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Community-CI: Mellanox Build Bot