bdev/nvme: avoid send_msg in bdev_nvme_enable_ctrlr
There is no need to make this operation async, as it is already async.
The bdev_nvme_reconnect_ctrlr_poll poller is registered below and polls
for reconnect completion.

Before send_msg with a controller as an argument, the ref should likely
be increased to prevent stale pointer usage when the message is
eventually executed. Instead using refcnt keep send_msg usage to a bare
minimum.

Change-Id: I05c789851e6d7c507df173be761ac13ff2a5aff7
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27075
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>