bdev/nvme: narrow critical section in nvme_ctrlr_depopulate_namespace
There is ref decr and checked in nvme_ctrlr_depopulate_namespace but it
is executed on the app thread already and it is being modified only on
the app thread by design.

While here also exclude nvme_ns->bdev clear from critical section. It is
not clear why it was there.

Change-Id: I0af11f5e963d8375a5f0ddc8623c010acb1de8cd
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27088
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>