bdev/nvme: remove stat from nvme_ns
It was introduced here
b317d8f39 bdev_nvme: add support to display io path stat of the NVMe bdev
recently there was a try to fix it here
d4cd02e71 bdev/nvme: Fix memory leak at shutdown when I/O path stat is enabled

But conceptually, the reason for these ns stat is questionable.
First, there was a thread race, as rpc_bdev_nvme_get_path_iostat was
reading nvme_ns->stat without taking the controller lock, while
nvme_ns->stat was being updated on IO threads. More importantly, the
race occurs because updates on IO threads happen at the time the path is
deleted. Path deletion is usually triggered by either namespace
depopulation or controller detachment/disconnect (followed by namespace
depopulation), so storing the statistics there (where the nvme_ns object
is about to be freed) races with rpc_bdev_nvme_get_path_iostat.

If the stats for a given path were meant to be saved, they should rather
be stored in the controller object, not the namespace object. However,
the RPC is about retrieving stats from the bdev per path (which might be
in the process of being released), not from the controller. Moreover,
even if one IO path is being deleted, it is likely that another will be
deleted shortly after — there is no good way to foresee and synchronize
that properly.

The results are also non-coherent for the reason mentioned in the
comment within _bdev_nvme_delete_io_path explaining that io_path stats
can still be updated afterwards.

Iterating over each channel of the nbdev should be sufficient, although
the issue still exists (marked with a TODO). One possible solution would
be to check the refcnt at the end of the channel iteration and fail
the RPC instead of returning non-coherent results.

This also seems to be the only reason for protecting the
ctrlr->namespaces list with a mutex, since all other flows that read,
write, or update this list are executed on the app thread.

Change-Id: I581ead93b3b720a04c1f105c75b536c97ea6b4a8
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27090
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Community-CI: Mellanox Build Bot