bdev: fix thread race in spdk_bdev_io_type_supported
The module’s .io_type_supported function was being invoked on IO
threads, while some of the objects e.g. in bdev/nvme module (nvme_ns,
ns, ctrlr) are modified (or even released) on the app thread causing
races and leading to undefined behavior.

The fix is to cache the supported IO types in the bdev itself during
spdk_bdev_register. This already runs on the app thread and is
considered thread safe.

There was a 4 byte hole in the spdk_bdev type in the first cache line,
so used it for the bitmask.

Fixes #3774

Change-Id: If7a98af8ff2bcc7b7a46a6e63b9297e3973553b4
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27143
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Pierre Lestringant <plestringant@kalrayinc.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>