test: skip some common and nvmf tests in no-huge mode
Not all test cases applies in the no-huge mode.
Either there is no support for the --no-huge command line option in the
underling tools being run or there is no real value in duplicating the
test in different modes of operation.
This patch disables common and nvmf functional tests whenever
SPDK_TEST_NO_HUGE is set.
Goal is to enable running the nvmf testing with the POSIX socket
implementation in the no-huge mode via regular autorun.sh job, like any
other job defined in the CI. Long term it can be expanded beyond the
nvmf scope. With no dependency on the hugepages it's possible to run the
test framework on regular GitHub VMs without the need to run nested
virtualization and/or on the containerized environments.

Change-Id: I87902327e0249ba2020a25320ca1edd7c52f1376
Signed-off-by: Krzysztof Goreczny <krzysztof.goreczny@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27010
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>