lib/bdev: Add API to get iostat by providing array of bdev names
In cases when a lot of bdevs are present inside application, it might
be useful to obtain a statistics over a concrete subset of bdevs.
Currently the only way to do it is to call bdev_get_iostat to get statistics
for all of bdevs and later filter only needed.
Patch introduces API to path bdev names array to bdev_get_iostat to obtain
statistics only for set bdevs.

Change-Id: I82de59d7439080985e34c8d73d15774537f35b5b
Signed-off-by: Vasilii Ivanov <iwanovvvasilij@gmail.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26942
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>