lib/blob: ensure ordering of operations on same extent_table_id
This patch is intended to cover the following scenario
in Issue #3682:
Modify the same extent page simultaneously.

Concurrent operations that update the same extent page can race
and lead to metadata corruption when their on-disk completion
order differs from the submission order.

This patch introduces a serialization mechanism for
concurrent updates to the same position in active.extent_pages.

Change-Id: Ia679c1d36a618aa5f5ef4c6d54ce2d2df7fee78e
Signed-off-by: Yalong Wang <yalong9@staff.sina.com.cn>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27041
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Community-CI: Mellanox Build Bot