python: add make build target
Previously the Makefile only exposed an `install` rule that invoked the
Python setup step on $(SPDK_ROOT_DIR)/python/, but there was no way to
build the package artifacts on their own.

Add a `build` rule that runs `python -m build` in the same directory,
providing a clean, explicit path to generate distribution artifacts
without triggering installation. This separates build and install
responsibilities and makes the workflow easier to script.

Change-Id: I55aa5a0791245bc06521afda74f859954a8dd968
Signed-off-by: Boris Glimcher <Boris.Glimcher@emc.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27058
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>