lib/nvmf: remove error paths from registrant list copying
namespace registrant list copying was written from the perspective of
iterating over the registrant list and copying into a hostid array of
fixed size.

Now the registrant list is capped at the max, it's no longer possible
for the two data structures to de-synchronize and there be more entries
in the list than SPDK_NVMF_MAX_NUM_REGISTRANTS.

This replaces the error handling with aborts(). The only way the list
could be longer the max is if memory corruption happens or developer
breaks the max enforcement.

Change-Id: I5322ed6a8e95e16568ae7585dc36a8eef30c75aa
Signed-off-by: Joel Cunningham <joel.cunningham@oracle.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27098
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Community-CI: Mellanox Build Bot