python: add make check target
Introduce a `check` target that depends on `build` and runs
`python -m twine check` on the artifacts in $(SPDK_ROOT_DIR)/python/dist/*.
Before this change, there was no way to verify the built distributions
from the Makefile.

Twine is the standard Python packaging tool used to validate and upload
distribution artifacts. Running its `check` command ensures the metadata
and structure of the package are sound before any upload or release.

Change-Id: I72acebf438bad642d0ef5967c175606bd7587307
Signed-off-by: Boris Glimcher <Boris.Glimcher@emc.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27107
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>