lib/nvmf: simplify poll group ns reservation update
The nvmf poll group namespace reservation update can be massively
simplified by not reusing nvmf_poll_group_update_subsystem() to update a
single namespace.

That function is a heavily handed approach to updating a single
namespace. It was meant for updating all nvmf poll group group memory
for an entire subsystems and is typically invoked via
nvmf_poll_group_add_subsystem() or nvmf_poll_group_remove_subsystem().

There are many failure paths in this function that could lead to
reservation state not being updated.

We can instead update only the namespace which has pending reservation
state updates. We know the poll group subsystem and namespace
data structures are valid when processing the reservation command so we
don't need all the checking either by the fact that IOs are only
processed when the subsystem and namesapce are in the active state.

Now it should not be possible for a poll group ns reservation update to
fail.  The spdk_for_each_channel() done callback now treats a non-zero
status as fatal.

subsystem_ut.c is updated to remove the mock version of
nvmf_poll_group_update_subsystem() and now we exercise all the nvmf poll
group update paths in all reservation unit tests which create poll
groups.

Change-Id: I3b5f8a0f0fbd5db367a92dbf588446df68288bd7
Signed-off-by: Joel Cunningham <joel.cunningham@oracle.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27247
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>