lib/nvmf: Move async event mask enum and nvmf_ctrlr_unmask_aen to header
This moves the nvmf_async_event_mask_bit enum from ctrlr.c to
nvmf_internal.h and also moves the nvmf_ctrlr_unmask_aen function to
the header so that these can be used elsewhere. This is a patch in
series of patches which makes the discovery log generation async.

Change-Id: Ic5925783c6cd06f2f2d8e0bcddd9f981fdbec189
Signed-off-by: Abhineet Pandey <abhineet.pandey@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27184
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Community-CI: Mellanox Build Bot