bdev/aio: unregister when device is detached
When the aio block device is detached from the system the aio poll
finds an error with value 0.
Whe using a file and the mount is removed then the value is -EIO.

Also when the aio bdev is rescanned, it fails with -EBUSY since
the dev size has reduced, to 0.

In these cases, we now check the block device size reported by the
backing fd. If the size is 0 then assume the device is detached and
we schedule bdev unregister in the app thread.

Change-Id: I45cb34105b45ef38221798dacd4cb179a2aa22b1
Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27193
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>