python: move dry run client class to package
dry run rpc client class was defined in scripts
which prevented the usage of it in the package
now moved to the correct location in the client.py
which is aligned with other client impl like golang.

now it can be re-used.

part of this changed, also moved tiny helper functions
- print nothing
- print deprecation

Change-Id: Ie25aeb793cf1b561845b5da07f12229d245e3456
Signed-off-by: Boris Glimcher <Boris.Glimcher@emc.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27200
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>