virtio/pci: map BARs on demand
We don't know upfront which BARs are used and which aren't.  If a BAR is
unused, it's size is zero, and if we still try to map it, we'll get
warnings from vtophys code trying to register it with IOMMU.

Instead, the BARs are mapped when they're referenced in virtio PCI
capabilities structure.

Change-Id: I0208b78ac08a4e9d7544a00595545a260474e169
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/28403
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Ben Walker <ben@nvidia.com>