lib/blob: spdk_bs_unload() will free bs on hotremoved disk
When the underlying disk is removed the I/O send to it will always return -EIO.
Make an exception for such case, to allow for cleaning up the blobstore.
Otherwise there is no other way to clean up the blobstore from memory.

Note that UT has no good way to check if the blobstore was freed, by clearing
the reference to it we rely on ASAN to catch any leaks.

Fixes #3727

Change-Id: I6aaa509b23f941de655866d274318f51911d75af
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27263
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>