accel/cuda: optimize completion of accel tasks
Task completion callbacks may take a bit of time if they do some tasks
or send new requests immediately in the callback code.
In such a scenario our cuda channels may be underloaded.

The patch moves completion of tasks to the end of the poller to allow
waiting requests to be started faster.

Change-Id: Ibac18d3a74c9d23954d6900e600de6b35c3c2e0a
Signed-off-by: Valerii Hlushkov <valerii.hlushkov@datacore.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27264
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>